tests/wpt/web-platform-tests/scroll-to-text-fragment/scroll-to-text-fragment.html

<!doctype html>
<title>Navigating to a text fragment directive</title>
<meta charset=utf-8>
<link rel="help" href="https://wicg.github.io/ScrollToTextFragment/">
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
let test_cases = [
  // Test non-text fragment directives
  {
    fragment: '#',
    expect_position: 'top',
    description: 'Empty hash'
  },
  {
    fragment: '#:~:text=this,is,test,page',
    expect_position: 'top',
    description: 'Text directive with invalid syntax, context terms without "-"'
  },
  {
    fragment: '#element:~:directive',
    expect_position: 'element',
    description: 'Generic fragment directive with existing element fragment'
  },
  // Test exact text matching, with all combinations of context terms
  {
    fragment: '#:~:text=test',
    expect_position: 'text',
    description:  'Exact text with no context'
  },
  {
    fragment: '#:~:text=this is a-,test',
    expect_position: 'text',
    description: 'Exact text with prefix'
  },
  {
    fragment: '#:~:text=test,-page',
    expect_position: 'text',
    description: 'Exact text with suffix'
  },
  {
    fragment: '#:~:text=this is a-,test,-page',
    expect_position: 'text',
    description: 'Exact text with prefix and suffix'
  },
  // Test text range matching, with all combinations of context terms
  {
    fragment: '#:~:text=this,page',
    expect_position: 'text',
    description: 'Text range with no context'
  },
  {
    fragment: '#:~:text=this-,is,test',
    expect_position: 'text',
    description: 'Text range with prefix'
  },
  {
    fragment: '#:~:text=this,test,-page',
    expect_position: 'text',
    description: 'Text range with suffix'
  },
  {
    fragment: '#:~:text=this-,is,test,-page',
    expect_position: 'text',
    description: 'Text range with prefix and suffix'
  },
  // Test partially non-matching text ranges
  {
    fragment: '#:~:text=this,none',
    expect_position: 'top',
    description: 'Text range with non-matching endText'
  },
  {
    fragment: '#:~:text=none,page',
    expect_position: 'top',
    description: 'Text range with non-matching startText'
  },
  // Test non-matching context terms
  {
    fragment: '#:~:text=this-,is,page,-none',
    expect_position: 'top',
    description: 'Text range with prefix and nonmatching suffix'
  },
  {
    fragment: '#:~:text=none-,this,test,-page',
    expect_position: 'top',
    description: 'Text range with nonmatching prefix and matching suffix'
  },
  // Test percent encoded characters
  {
    fragment: '#:~:text=this%20is%20a%20test%20page',
    expect_position: 'text',
    description: 'Exact text with percent encoded spaces'
  },
  {
    fragment: '#:~:text=test%20pag',
    expect_position: 'top',
    description: 'Non-whole-word exact text with spaces'
  },
  {
    fragment: '#:~:text=%26%2C%2D',
    expect_position: 'text',
    description: 'Fragment directive with percent encoded syntactical characters "&,-"'
  },
  {
    fragment: '#:~:text=%2665',
    expect_position: 'text',
    description: 'Fragment directive with percent encoded non-ASCII unicode character'
  },
  {
    fragment: '#:~:text=!$\'()*+./:;=?@_~',
    expect_position: 'text',
    description: 'Fragment directive with all TextMatchChars'
  },
  // Test multiple text directives
  {
    fragment: '#:~:text=this&text=test,page',
    expect_position: 'text',
    description: 'Multiple matching exact texts'
  },
  {
    fragment: '#:~:text=tes&text=age',
    expect_position: 'top',
    description: 'Multiple non-whole-word exact texts'
  },
  // Test text directive behavior when there's an element fragment identifier
  {
    fragment: '#element:~:text=test',
    expect_position: 'text',
    description: 'Text directive with existing element fragment'
  },
  {
    fragment: '#pagestate:~:text=test',
    expect_position: 'text',
    description: 'Text directive with nonexistent element fragment'
  },
  {
    fragment: '#element:~:text=nomatch',
    expect_position: 'element',
    description: 'Non-matching text directive with existing element fragment'
  },
  {
    fragment: '#pagestate:~:text=nomatch',
    expect_position: 'top',
    description: 'Non-matching text directive with nonexistent element fragment'
  },
  // Test ambiguous text matches disambiguated by context terms
  {
    fragment: '#:~:text=more-,test%20page',
    expect_position: 'more-text',
    description: 'Multiple match text directive disambiguated by prefix'
  },
  {
    fragment: '#:~:text=test%20page,-text',
    expect_position: 'more-text',
    description: 'Multiple match text directive disambiguated by suffix'
  },
  {
    fragment: '#:~:text=more-,test%20page,-text',
    expect_position: 'more-text',
    description: 'Multiple match text directive disambiguated by prefix and suffix'
  },
  // Test context terms separated by node boundaries
  {
    fragment: '#:~:text=prefix-,test%20page,-suffix',
    expect_position: 'cross-node-context',
    description: 'Text directive with context terms separated by node boundaries'
  },
];

for (const test_case of test_cases) {
  promise_test(t => new Promise(resolve => {
    let channel = new BroadcastChannel('scroll-to-text-fragment');
    channel.addEventListener("message", e => {
      resolve(e.data);
    }, {once: true});

    test_driver.bless('Open a URL with a text fragment directive', () => {
      window.open('scroll-to-text-fragment-target.html' + test_case.fragment, '_blank', 'noopener');
    });
  }).then(data => {
    assert_equals(data.href.indexOf(':~:'), -1, 'Expected fragment directive to be stripped from the URL.');
    assert_equals(data.scrollPosition, test_case.expect_position,
                  `Expected ${test_case.fragment} (${test_case.description}) to scroll to ${test_case.expect_position}.`);
  }), `Test navigation with fragment: ${test_case.description}`);
}

promise_test(t => new Promise(resolve => {
  let channel = new BroadcastChannel('scroll-to-text-fragment');
  channel.addEventListener("message", e => {
    assert_equals(e.data.href.indexOf(':~:'), -1, 'Expected fragment directive to be stripped from the URL.');

    // The first navigation has no user activation.
    assert_equals(e.data.scrollPosition, 'top', 'Expected window.open() with no user activation to not activate text fragment directive.');

    // Now ensure that a navigation with a user activation does activate the text fragment directive.
    test_driver.bless('Open a URL with a text fragment directive', () => {
      window.open('scroll-to-text-fragment-target.html#:~:text=test', '_blank', 'noopener');
    });
    channel.addEventListener("message", e => {
      resolve(e.data.scrollPosition);
    }, {once: true});
  }, {once: true});

  window.open('scroll-to-text-fragment-target.html#:~:text=test', '_blank', 'noopener');
}).then(scrollPosition => {
  assert_equals(scrollPosition, 'text', 'Expected window.open() with a user activation to scroll to text.');
}), 'Test that a text fragment directive is not activated without a user activation');

promise_test(t => new Promise(resolve => {
  let channel = new BroadcastChannel('scroll-to-text-fragment');
  channel.addEventListener("message", e => {
    assert_equals(e.data.href.indexOf(':~:'), -1, 'Expected fragment directive to be stripped from the URL.');

    // The first navigation has an opener.
    assert_equals(e.data.scrollPosition, 'top', 'Expected window.open() with opener to not activate text fragment directive.');

    // Now ensure that a navigation with noopener does activate the text fragment directive.
    test_driver.bless('Open a URL with a text fragment directive', () => {
      window.open('scroll-to-text-fragment-target.html#:~:text=test', '_blank', 'noopener');
    });
    channel.addEventListener("message", e => {
      resolve(e.data.scrollPosition);
    }, {once: true});
  }, {once: true});

  test_driver.bless('Open a URL with a text fragment directive', () => {
    window.open('scroll-to-text-fragment-target.html#:~:text=test', '_blank');
  });
}).then(scrollPosition => {
  assert_equals(scrollPosition, 'text', 'Expected window.open() with noopener to scroll to text.');
}), 'Test that a text fragment directive is not activated when there is a window opener.');

promise_test(t => new Promise(resolve => {
  let channel = new BroadcastChannel('scroll-to-text-fragment');
  channel.addEventListener("message", e => {
    resolve(e.data);
  }, {once: true});

  let frame = document.createElement('iframe');
  document.body.appendChild(frame);

  test_driver.bless('Navigate the iframe with a text fragment directive', () => {
    frame.src = 'scroll-to-text-fragment-target.html#:~:text=test';
  });
}).then(data => {
  assert_equals(data.href.indexOf(':~:'), -1, 'Expected fragment directive to be stripped from the URL.');
  assert_equals(data.scrollPosition, 'top', 'Expected iframe navigation to not activate text fragment directive.');
}), 'Test that a text fragment directive is not activated within an iframe.');
</script>
